pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["Leo <foxes97@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
django-environ = "^0.11.0"
djangorestframework = "^3.15"
django-cors-headers = "^4.3"
djangorestframework-simplejwt = "^5.3"
djoser = "^2.2"
django-filter = "^24.3"
drf-spectacular = "^0.27"
psycopg2-binary = "^2.9"
social-auth-app-django = "^5.6.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
django-rest-framework-social-oauth2 = "^1.2.0"
requests = "^2.32.5"
dataclasses-json = "^0.6.7"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"